Added obstacles_file - works fine
