+++ main.py
@@ -55,68 +55,11 @@ class Player:
 		win.blit(self.img, (self.x,self.y))
 		self.runCount += 1 
 
-class Tree:
-	"""
-	Describes tree objects. Currently it contains 11 types of trees. It contains the x position, y position of tree.
-	Also contains a draw method to draw trees onto the screen. The type of tree that is to be created can be determined using num variable.
-	"""
-	# Loading images 
-	num_of_imgs = 11
-	imgs = [pygame.image.load(os.path.join('Utils/Pics/Obstacles/', "tree"+ str(x) + '.png')).convert_alpha() for x in range(0, num_of_imgs)]
-	resized_imgs = [pygame.transform.scale(img, (int(img.get_width()//1.2), int(img.get_height()//1.2))) for img in imgs]
-
-	obstacles = []
-
-
-	def __init__(self, x, num):
-		self.num = num
-		self.img = self.resized_imgs[self.num]
-		self.x = x
-		self.y = ground_y - self.img.get_height() + 230 # push the image slightly down
-	
-		# Width and height of obstacle
-		self.width = self.img.get_width()
-		self.height = self.img.get_height()
-
-	# Draws the obstacle onto the screen
-	def draw(self,win):
-		win.blit(self.img, (self.x, self.y)) 
-
-class Other_obstacles:
-	"""
-	Describes other obstacle objects. Contains rocks and shrubs. It contains the x position, y position. The various obstacles are created using num variable.
-	Also contains a draw method to draw the tree onto the screen.
-	"""
-	# Loading images 
-	num_of_imgs = 2
-	imgs = [pygame.image.load(os.path.join('Utils/Pics/Obstacles/', "obstacle"+ str(x) + '.png')).convert_alpha() for x in range(0, num_of_imgs)]
-	resized_imgs = [pygame.transform.scale(img, (int(img.get_width()), int(img.get_height()))) for img in imgs]
-
-	obstacles = []
-
-	def __init__(self, x, num):
-		self.num = num
-		self.img = self.resized_imgs[self.num]
-		self.x = x
-		self.y = ground_y - self.img.get_height() + 250 # push the image slightly down
-
-		# Width and height of obstacle
-		self.width = self.img.get_width()
-		self.height = self.img.get_height()
-
-	# Draws the obstacle onto the screen
-	def draw(self,win):
-		win.blit(self.img, (self.x, self.y))
-
-obstacle_classes = [Tree, Other_obstacles]
-
-# HELPER FUNCTIONS
-
 def draw_scene_and_obstacles():
 	"""
 	Draws the background, foreground, obstacles and coins.
 	"""
-	global bg_x, bg_width, ground_x, ground_width
+	global bg_x, bg_width, ground_x, ground_y, ground_width
 	# Drawing background
 	win.blit(bg, (bg_x, 0))
 	win.blit(bg, (bg_width,0))
@@ -130,16 +73,6 @@ def draw_scene_and_obstacles():
 		bg_x = bg.get_width()
 	if bg_width < bg.get_width() * -1:
 		bg_width = bg.get_width()
-
-	# Drawing obstacles
-	for obstacle in Other_obstacles.obstacles:
-		obstacle.draw(win)
-	for obstacle in Tree.obstacles:
-		obstacle.draw(win)
-
-	# Draw coins
-	for coin in coins_file.Coin.coins_list:
-		coin.draw(win)
 
 	# Drawing ground
 	win.blit(ground, (ground_x, ground_y))
@@ -156,53 +89,6 @@ def draw_scene_and_obstacles():
 		ground_width = (flipped_ground.get_width()-5)
 
 	return bg_x, bg_width, ground_x, ground_width
-
-
-def create_tree_obstacle():
-	"""
-	Creates a random tree obstacle.
-	"""
-	random_num = random.randrange(0,Tree.num_of_imgs)       	# range over the number of obstacles
-	random_x = random.randint(bg.get_width(), bg.get_width()+500)   # random inital x position of obstacle
-	# Add new obstacle to obstacles list
-	Tree.obstacles.append(Tree(random_x,random_num))
-
-def create_other_obstacle():
-	"""
-	Creates a random obstacle.
-	"""
-	random_num = random.randrange(0,Other_obstacles.num_of_imgs)       	# range over the number of obstacles
-	random_x = random.randint(bg.get_width(), bg.get_width()+500)   # random inital x position of obstacle
-	# Add new obstacle to obstacles list
-	Other_obstacles.obstacles.append(Other_obstacles(random_x,random_num))
-
-def update_obstacle_position():
-	"""
-	Creates the moving effect. If obstacle goes offscreen, removes it from obstacles list"
-	"""
-	for element in obstacle_classes:
-		for obstacle in element.obstacles:
-			if obstacle.x < -1*obstacle.width:
-				element.obstacles.remove(obstacle)
-			else:
-				obstacle.x -= foreground_speed
-
-def collision_with_obstacle():
-	"""
-	Pixel perfect collision implemented using masks. Checks for collision of mask of player against other
-	obstacles only if the obstacle is near to the player. Returns True if collision occurred, else False.
-	"""
-	player_mask = pygame.mask.from_surface(player.img)
-	for element in obstacle_classes:
-		for obstacle in element.obstacles:
-			if obstacle.x < (player.x + player.img.get_width()+10):	# Checking for collision if near player
-				obstacle_mask = pygame.mask.from_surface(obstacle.img)
-				offset = obstacle.x - player.x, obstacle.y - player.y
-				boolean = player_mask.overlap(obstacle_mask, offset)
-				if boolean:
-					return True
-	return False
-
 
 def display_mouse_pointer_coordinates(mx,my):
 	"""
@@ -226,6 +112,7 @@ def display_collision_message():
 
 if __name__ == '__main__':
 	import coins_file
+	import obstacles_file
 
 	player = Player(250, 313)	# Creating an instance of the class Player
 
@@ -250,6 +137,10 @@ if __name__ == '__main__':
 	while run:
 		# Draws stuff to be displayed in window
 		draw_scene_and_obstacles()
+		obstacles_file.draw_obstacles(win)
+		# Draw coins
+		for coin in coins_file.Coin.coins_list:
+			coin.draw(win)
 		
 		# Get mouse pointer coordinates
 		(mx, my) = pygame.mouse.get_pos()
@@ -283,11 +174,11 @@ if __name__ == '__main__':
 			if event.type == pygame.USEREVENT+1:
 				coins_file.create_coin()
 			if event.type == pygame.USEREVENT+2:
-				create_tree_obstacle()
+				obstacles_file.create_tree_obstacle()
 			if event.type == pygame.USEREVENT+3:
-				create_other_obstacle()
+				obstacles_file.create_other_obstacle()
 		
-		update_obstacle_position()
+		obstacles_file.update_obstacle_position()
 		coins_file.update_coins_position()
 
 		# Coin collection
@@ -297,7 +188,7 @@ if __name__ == '__main__':
 		coins_file.display_num_coins_collected(win)
 
 		# Collision with Obstacles
-		collision_occured = collision_with_obstacle()	# Checks collision and returns bool 
+		collision_occured = obstacles_file.collision_with_obstacle(player)	# Checks collision and returns bool 
 		if collision_occured:		# Dummy exit
 			display_collision_message()
 			pygame.display.update()
